separate industry filter
