+++ mod/filter_industry/set/right/browse_company_filter/industry_filter.rb
@@ -1,0 +1,3 @@
+def filter_keys
+  super << :industry
+end
