+++ mod/filter_industry/set/right/metric_company_filter/industry_filter.rb
@@ -1,0 +1,3 @@
+def advanced_filter_keys
+  super.insert 1, :industry
+end
